Fix the bugs of the multi-pane mode
Now even the tablet layout works.

Bugs still to consider, when screen is rotated the tablet layout loses
its detailed fragment.

Other things to fix, renaming and deleting should update the detailed
view (I should check when it does and if it doesn't I should fix it),
still need to test it